+++ dart/appwrouter/lib/src/utils/src/redirect.dart
@@ -1,0 +1,5 @@
+/// Redirects to the specified path.
+void redirect(dynamic req, {required String path}) {
+  // ignore: avoid_dynamic_calls
+  req.path = path;
+}