+++ dart/appwrouter/CHANGELOG.md
@@ -1,3 +1,32 @@
+# 0.2.5
+
+- fix: Class 'RuntimeResponse' has no instance getter 'html'.
+
+# 0.2.4
+
+- fix: cannot parse params
+
+# 0.2.3
+
+- feat: added AppwrouterResponse and AppwrouterRequest to mimic the request and response object in the Appwrite SDK
+
+# 0.2.2
+
+- fix: fix the dart environment to compatible with appwrite dart 3.0.0 runtime
+
+# 0.2.1
+
+- fix: fix the package for appwrite
+
+# 0.2.0
+
+- feat: Added utilities for redirect
+- refactor: Change the FutureOr into Future
+
+# 0.1.3
+
+- refactor: Change the FutureOr into Future
+
 # 0.1.2
 
 - feat: improve documentation